Change light type

+++ color.hpp
@@ -6,9 +6,9 @@
 class Color
 {
 public:
-    short RED;
-    short GREEN;
-    short BLUE;
+    uint8_t RED;
+    uint8_t GREEN;
+    uint8_t BLUE;
 };
 
 std::ostream& operator<<(std::ostream &out , const Color &color);
